db updating isn't working correctly
This is the result of carrying the problem by hand
by creating explicit Payment.csv and Session.csv files
from the most recent CarolinaCardClub.db as left on
the local server.  That file was not correctly backed
up to the web server, however, so those files had to
be extracted and then used to update an active local
copy of the db -- used sqlite for that, but sheesh.